requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
httpx==0.27.0
feedparser==6.0.11
trafilatura==1.7.0
python-dateutil==2.9.0.post0
langdetect==1.0.9
yake==0.4.8
simhash==2.1.2
spacy==3.7.4
lxml==5.2.2
httpx 
tldextract==5.1.2
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
hdbscan==0.8.37
umap-learn==0.5.6
bertopic==0.16.4
sentence-transformers==3.0.1
redis==5.0.4
prometheus-client==0.20.0
